fabric member session moved to fabric package
