+++ packages/core/src/rooms/index.ts
@@ -12,4 +12,3 @@ export * from './methods'
 export * from './RoomSessionRecording'
 export * from './RoomSessionPlayback'
 export * from './RoomSessionStream'
-export * from './RoomSessionMember'
